+++ main/events/guildDelete.js
@@ -7,6 +7,7 @@ module.exports = async (bot, guild) => {
   //Update bot-data.json
   let botdata = require("../../data/global/bot-data.json");
   botdata.stats.totalGuilds = bot.guilds.cache.size;
+  botdata.stats.totalUsers = bot.guilds.cache.reduce((a, g) => a + g.memberCount, 0);
   fs.writeFileSync(`./data/global/bot-data.json`, JSON.stringify(botdata, null, 4), (err) => {
     if (err) return bot.log.post("error", err);
   });